Define URIReference, redefine IndirectValue
Previous distinction between IndirectValue#as
and #get weren't making sense to me, so I rewrote
the description of #as to indicate that it is
just 'converting to a different representation'
(with the implication that 'conversion' will often
be returning `(T)this`), and less emphasis on
that returned value being a different representation
of the indirect value itself rather than of the
represented concept.  My current thinking being that
anything you might return is a representation,
it's just a matter of agreeing on what is being
represented and how, that's what the tag is for,
and that therefore there is no point in trying
to make distinctions between different 'levels'
of representation beyond what the existence of
IndirectValue already gives you.

- IndirectValue :: tells you exactly how to interpret it
  so long as you know what to as() for
- Any other Java object :: interpret values as meaning
  'themselves' in the conventional way (an Integer
  is a number, a String is a sequence of characters,
  etc).

...with the caveat that 'the conventional way' is context-sensitive.