Continuations, returnTo
Remove ScriptInterpreter because it wasn't used.

Implement the read-execute loop using a return
continuation which is implicitly used when running
off the end of the current list of operations instead
of adding a jump to the end of to-be-immediately-executed
op lists.  This seems like it might actually be slightly
wrong, since normally a 'return' in that position
would quit the program.

TODO: Fix later.